+++ src/lib_media/demux/gpac_demux_mp4_full.cpp
@@ -134,8 +134,8 @@ class GPACDemuxMP4Full : public ModuleS {
 				auto const DTSOffset = reader.movie->getDTSOffset(FIRST_TRACK);
 				//here we dump some sample info: samp->data, samp->dataLength, samp->isRAP, samp->DTS, samp->CTS_Offset
 				m_host->log(Debug, format("Found sample #%s(#%s) of length %s , RAP: %s, DTS: %s, CTS: %s",
-				        reader.sampleIndex, sample->dataLength,
-				        sample->IsRAP, sample->DTS + DTSOffset, sample->DTS + DTSOffset + sample->CTS_Offset).c_str());
+				        reader.sampleIndex, sample->dataLength, (int)sample->IsRAP,
+				        sample->DTS + DTSOffset, sample->DTS + DTSOffset + sample->CTS_Offset).c_str());
 				reader.sampleIndex++;
 
 				auto out = output->allocData<DataRaw>(sample->dataLength);
